move to vite plugin
