+++ packages/plugin-connector/package.json
@@ -24,7 +24,8 @@
 		"svelte-preprocess": "^5.0.3",
 		"sveltekit-autoimport": "^1.7.0",
 		"yaml": "^2.2.2",
-		"zod": "^3.21.4"
+		"zod": "^3.21.4",
+		"chokidar": "^3.5.3"
 	},
 	"devDependencies": {
 		"@parcel/core": "^2.8.3",
