+++ packages/plugin-connector/src/build-plugins/vite.js
@@ -1,0 +1,6 @@
+import { sourceQueryHMR } from './source-query-hmr.js';
+
+/** @type {() => import("vite").UserConfig["plugins"]} */
+export const evidenceVitePlugin = () => {
+	return [sourceQueryHMR()];
+};
