+++ packages/plugin-connector/src/index.js
@@ -1,3 +1,4 @@
 export * from './build-plugins/preprocess';
 export * from './build-plugins/rollup';
+export * from './build-plugins/vite';
 export * from './data-sources';
